% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Admission}
\alias{Admission}
\title{Graduate school admissions data}
\format{A data set with 400 rows and 4 variables.}
\source{
UCLA IDRE \url{http://stats.idre.ucla.edu/r/dae/logit-regression/}
}
\usage{
Admission
}
\description{
A data set containing 400 graduate school admissions decisions.
}
\keyword{datasets}
